+++ apps/member-profile/app/routes/_profile.directory.tsx
@@ -14,6 +14,7 @@ import { useState } from 'react';
 import { Filter, Plus } from 'react-feather';
 import { match } from 'ts-pattern';
 
+import { SchoolCombobox } from '@oyster/core/education.ui';
 import { CityCombobox } from '@oyster/core/location.ui';
 import { db } from '@oyster/db';
 import { type ExtractValue } from '@oyster/types';
@@ -35,7 +36,6 @@ import { listMembersInDirectory } from '@/member-profile.server';
 import {
   ListMembersInDirectoryWhere,
   ListSearchParams,
-  SchoolCombobox,
 } from '@/member-profile.ui';
 import { CompanyCombobox } from '@/shared/components/company-combobox';
 import { EthnicityCombobox } from '@/shared/components/ethnicity-combobox';
