+++ apps/member-profile/app/shared/components/education-form.tsx
@@ -5,7 +5,7 @@ import React, {
   useState,
 } from 'react';
 
-import { MajorCombobox } from '@oyster/core/education.ui';
+import { MajorCombobox, SchoolCombobox } from '@oyster/core/education.ui';
 import { type Major } from '@oyster/types';
 import { DatePicker, Form, Input, Select } from '@oyster/ui';
 import { toTitleCase } from '@oyster/utils';
@@ -14,7 +14,6 @@ import {
   DegreeType,
   FORMATTED_DEGREEE_TYPE,
   type School,
-  SchoolCombobox,
 } from '@/member-profile.ui';
 
 const EducationFormContext = React.createContext({
