+++ packages/core/src/modules/education/education.ui.tsx
@@ -1,4 +1,5 @@
-import { useState } from 'react';
+import { useFetcher } from '@remix-run/react';
+import React, { useEffect, useState } from 'react';
 
 import { Major } from '@oyster/types';
 import {
@@ -11,6 +12,7 @@ import {
   Form,
   Input,
   type InputProps,
+  useDelayedValue,
 } from '@oyster/ui';
 import { toEscapedString, toTitleCase } from '@oyster/utils';
 
@@ -173,6 +175,79 @@ export function MajorCombobox({
               </ComboboxItem>
             );
           })}
+        </ul>
+      </ComboboxPopover>
+    </Combobox>
+  );
+}
+
+// School Combobox
+
+type SchoolComboboxProps = Pick<InputProps, 'readOnly' | 'required'> & {
+  defaultValue?: Pick<School, 'id' | 'name'>;
+  name: string;
+  onSelect?: React.HTMLProps<HTMLButtonElement>['onClick'];
+  popoverProps?: Partial<React.ComponentProps<typeof ComboboxPopover>>;
+};
+
+export function SchoolCombobox({
+  defaultValue = { id: '', name: '' },
+  name,
+  onSelect,
+  popoverProps = {},
+  readOnly,
+  required,
+}: SchoolComboboxProps) {
+  const [search, setSearch] = useState<string>('');
+
+  const delayedSearch = useDelayedValue(search, 250);
+
+  const fetcher = useFetcher<{
+    schools: Pick<School, 'id' | 'name'>[];
+  }>();
+
+  useEffect(() => {
+    fetcher.submit(
+      { search: delayedSearch },
+      {
+        action: '/schools',
+        method: 'get',
+      }
+    );
+  }, [delayedSearch]);
+
+  const schools = fetcher.data?.schools || [];
+
+  return (
+    <Combobox
+      defaultDisplayValue={defaultValue.name}
+      defaultValue={defaultValue.id}
+    >
+      <ComboboxInput
+        id={name}
+        name={name}
+        onChange={(e) => setSearch(e.currentTarget.value)}
+        readOnly={readOnly}
+        required={required}
+      />
+
+      <ComboboxPopover {...popoverProps}>
+        <ul>
+          {schools.map((school) => {
+            return (
+              <ComboboxItem
+                key={school.id}
+                onSelect={onSelect}
+                value={school.id}
+              >
+                {school.name}
+              </ComboboxItem>
+            );
+          })}
+
+          <ComboboxItem onSelect={onSelect} value="other">
+            Other
+          </ComboboxItem>
         </ul>
       </ComboboxPopover>
     </Combobox>
